Defer "pure expression .. in statement position" warnings until refchecks.
This allows a macro to transform code that used to warn into something that
does not.
