+++ test/files/neg/macro-invalidret.check
@@ -27,8 +27,5 @@ java.lang.NullPointerException
 Macros_Test_2.scala:15: error: macro implementation is missing
   foo4
   ^
-Macros_Test_2.scala:17: warning: a pure expression does nothing in statement position; multiline expressions may require enclosing parentheses
-  foo6
-  ^
-two warnings found
+one warning found
 5 errors found
