+++ test/files/neg/t1181.check
@@ -3,8 +3,4 @@ t1181.scala:9: error: type mismatch;
  required: Symbol
      _ => buildMap(map.updated(keyList.head, valueList.head), keyList.tail, valueList.tail)
                   ^
-t1181.scala:8: warning: a pure expression does nothing in statement position; multiline expressions might require enclosing parentheses
-     case (Nil, Nil) => map
-                        ^
-one warning found
 one error found
