+++ test/files/neg/unit-returns-value.check
@@ -1,9 +1,9 @@
-unit-returns-value.scala:4: warning: a pure expression does nothing in statement position
-    if (b) return 5
-                  ^
 unit-returns-value.scala:4: warning: enclosing method f has result type Unit: return value discarded
     if (b) return 5
            ^
+unit-returns-value.scala:4: warning: a pure expression does nothing in statement position
+    if (b) return 5
+                  ^
 unit-returns-value.scala:22: warning: a pure expression does nothing in statement position; multiline expressions might require enclosing parentheses
     i1 // warn
     ^
