+++ test/files/run/reify_lazyunit.check
@@ -1,6 +1,3 @@
-reify_lazyunit.scala:6: warning: a pure expression does nothing in statement position; multiline expressions might require enclosing parentheses
-    lazy val x = { 0; println("12")}
-                   ^
 12
 one
 two
