+++ test/files/run/t3488.check
@@ -1,13 +1,13 @@
-t3488.scala:4: warning: a pure expression does nothing in statement position
-  println(foo { val List(_*)=List(0); 1 } ())
-                                      ^
 t3488.scala:4: warning: multiline expressions might require enclosing parentheses; a value can be silently discarded when Unit is expected
   println(foo { val List(_*)=List(0); 1 } ())
                                       ^
-t3488.scala:5: warning: a pure expression does nothing in statement position
-  println(foo { val List(_*)=List(0); 1 } (1))
+t3488.scala:4: warning: a pure expression does nothing in statement position
+  println(foo { val List(_*)=List(0); 1 } ())
                                       ^
 t3488.scala:5: warning: multiline expressions might require enclosing parentheses; a value can be silently discarded when Unit is expected
+  println(foo { val List(_*)=List(0); 1 } (1))
+                                      ^
+t3488.scala:5: warning: a pure expression does nothing in statement position
   println(foo { val List(_*)=List(0); 1 } (1))
                                       ^
 0
